+++ R/merge_rrm_on_vri.R
@@ -1,21 +1,32 @@
-#' merge_rrm_on_vri_bear
+#' merge_rrm_on_vri
 #'
 #' merge suitability and capability rating from rrm onto vri-bem
 #' and calculate highest value and weighted average for each scores
 #'
 #' @param vri_bem sf object that represent VRI (vegetation ressource inventory) features
 #' @param rrm_dt data.table object that contains the rrm
+#' @param animal character, "bear" or "moose"
 #' @param return_sf logical, if TRUE  return sf object , if FALSE return data.table object and update by reference
 #' @return sf  vri-bem object with new columns for rating
 #' @import data.table
 #' @export
-merge_rrm_on_vri_bear <- function(vri_bem, rrm_dt, return_sf = TRUE) {
+merge_rrm_on_vri <- function(vri_bem, rrm_dt, animal, return_sf = TRUE) {
 
   setDT(vri_bem)
 
   # calc capability rating and format rrm table
   calc_capability_rating(rrm_dt = rrm_dt)
   format_rrm_dt(rrm_dt = rrm_dt)
+
+  rating_variables <- grep("_6C$", names(rrm_dt), value = T)
+  rating_variables_no_suffix <- substr(rating_variables, 1, nchar(rating_variables) - 3)
+  cap_rating_variables <- paste0(rating_variables_no_suffix, "_CAP")
+
+  variables_merge_expr <- paste0("list(", paste(paste0("", c(rating_variables, cap_rating_variables)), collapse = ","), ")")
+
+  first_decile_variables <- c(paste0(rating_variables_no_suffix, "_SU_1"), paste0(cap_rating_variables, "_1"))
+  second_decile_variables <- c(paste0(rating_variables_no_suffix, "_SU_2"), paste0(cap_rating_variables, "_2"))
+  third_decile_variables <- c(paste0(rating_variables_no_suffix, "_SU_3"), paste0(cap_rating_variables, "_3"))
 
   # Create temp Stand variable to match variable used in the creation of the rrm
   vri_bem[ , STAND_A1_temp := STAND_A1]
@@ -30,108 +41,130 @@ merge_rrm_on_vri_bear <- function(vri_bem, rrm_dt, return_sf = TRUE) {
   which_not_in_rrm_3 <- which(vri_bem[["STAND_A3"]] %in% c("B", "C", "M") & !vri_bem[["STRCT_S3"]] %in% c("4", "5", "6", "7"))
   set(vri_bem, i = which_not_in_rrm_3, j = "STAND_A3_temp", value = "")
 
-  # merge on decile 1 ----
-  vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
-                         BGC_PHASE = Bgc_phase, BEUMC_S1 = Beumc, SLOPE_MOD = Slope_mod,
-                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
-                         CROWN_BEAR_1 = Crown_Bear, STRCT_S1 = Strct_d, STAND_A1_temp = Stand_d),
-          c("MURAR_PEFD_SU_1", "MURAR_PEFD_CAP_1") := .(i.MURAR_PEFD_6C, i.MURAR_PEFD_CAP)]
+  if (animal == "bear") {
+    merge_rating_bear(vri_bem = vri_bem,
+                      rrm_dt = rrm_dt,
+                      rating_variables = list(first_decile_variables,
+                                              second_decile_variables,
+                                              third_decile_variables),
+                      rating_variables_expr = variables_merge_expr)
+  }
 
-  set(vri_bem, i = which_not_in_rrm_1, j = "MURAR_PEFD_SU_1", value = NA)
-  # merge on decile 2 ----
-  vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
-                         BGC_PHASE = Bgc_phase, BEUMC_S2 = Beumc, SLOPE_MOD = Slope_mod,
-                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
-                         CROWN_BEAR_2 = Crown_Bear, STRCT_S2 = Strct_d, STAND_A2_temp = Stand_d),
-          c("MURAR_PEFD_SU_2", "MURAR_PEFD_CAP_2") := .(i.MURAR_PEFD_6C, i.MURAR_PEFD_CAP)]
+  if (animal == "moose") {
+    merge_rating_moose(vri_bem = vri_bem,
+                      rrm_dt = rrm_dt,
+                      rating_variables = list(first_decile_variables,
+                                              second_decile_variables,
+                                              third_decile_variables),
+                      rating_variables_expr = variables_merge_expr)
+  }
 
-  set(vri_bem, i = which_not_in_rrm_2, j = "MURAR_PEFD_SU_2", value = NA)
-
-  # merge on decile 3 ----
-  vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
-                         BGC_PHASE = Bgc_phase, BEUMC_S3 = Beumc, SLOPE_MOD = Slope_mod,
-                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
-                         CROWN_BEAR_3 = Crown_Bear, STRCT_S3 = Strct_d, STAND_A3_temp = Stand_d),
-          c("MURAR_PEFD_SU_3", "MURAR_PEFD_CAP_3") := .(i.MURAR_PEFD_6C, i.MURAR_PEFD_CAP)]
-
-  set(vri_bem, i = which_not_in_rrm_3, j = "MURAR_PEFD_SU_3", value = NA)
+  set(vri_bem, i = which_not_in_rrm_1, j = paste0(rating_variables_no_suffix, "_SU_1"), value = NA)
+  set(vri_bem, i = which_not_in_rrm_2, j = paste0(rating_variables_no_suffix, "_SU_2"), value = NA)
+  set(vri_bem, i = which_not_in_rrm_3, j = paste0(rating_variables_no_suffix, "_SU_3"), value = NA)
 
   # calc highest suitability value ----
 
-  # assign temporary worst rating  to rating NA to make calculation of best rating easier
+  which_na_list <- list()
+  for (suitability_variable in rating_variables_no_suffix) {
 
-  set(vri_bem, i = which((vri_bem[["FORESTED_1"]] == "Y" & vri_bem[["STRCT_S1"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_1"]] == "N" & vri_bem[["STRCT_S1"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = "MURAR_PEFD_SU_1", value = NA)
-  set(vri_bem, i = which((vri_bem[["FORESTED_2"]] == "Y" & vri_bem[["STRCT_S2"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_2"]] == "N" & vri_bem[["STRCT_S2"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = "MURAR_PEFD_SU_2", value = NA)
-  set(vri_bem, i = which((vri_bem[["FORESTED_3"]] == "Y" & vri_bem[["STRCT_S3"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_3"]] == "N" & vri_bem[["STRCT_S3"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = "MURAR_PEFD_SU_3", value = NA)
+    first_suit_var <- paste0(suitability_variable, "_SU_1")
+    second_suit_var <- paste0(suitability_variable, "_SU_2")
+    third_suit_var <- paste0(suitability_variable, "_SU_3")
+    high_value_suit_var <- paste0(suitability_variable, "_SU_HV")
+    weighted_average_suit_var <- paste0(suitability_variable, "_SU_WA")
 
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_SU_1"]] > 6), j = "MURAR_PEFD_SU_1", value = NA)
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_SU_2"]] > 6), j = "MURAR_PEFD_SU_2", value = NA)
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_SU_3"]] > 6), j = "MURAR_PEFD_SU_3", value = NA)
+    # assign temporary worst rating  to rating NA to make calculation of best rating easier
+    set(vri_bem, i = which((vri_bem[["FORESTED_1"]] == "Y" & vri_bem[["STRCT_S1"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_1"]] == "N" & vri_bem[["STRCT_S1"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = first_suit_var, value = NA)
+    set(vri_bem, i = which((vri_bem[["FORESTED_2"]] == "Y" & vri_bem[["STRCT_S2"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_2"]] == "N" & vri_bem[["STRCT_S2"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = second_suit_var, value = NA)
+    set(vri_bem, i = which((vri_bem[["FORESTED_3"]] == "Y" & vri_bem[["STRCT_S3"]] == "7a" & vri_bem[["VRI_AGE_CL_STS"]] == -1) | (vri_bem[["FORESTED_3"]] == "N" & vri_bem[["STRCT_S3"]] == "" & vri_bem[["ABOVE_ELEV"]] == "N")), j = third_suit_var, value = NA)
 
-  which_su_1_is_na <- which(is.na(vri_bem[["MURAR_PEFD_SU_1"]]))
-  which_su_2_is_na <- which(is.na(vri_bem[["MURAR_PEFD_SU_2"]]))
-  which_su_3_is_na <- which(is.na(vri_bem[["MURAR_PEFD_SU_3"]]))
+    set(vri_bem, i = which(vri_bem[[first_suit_var]] > 6), j = first_suit_var, value = NA)
+    set(vri_bem, i = which(vri_bem[[second_suit_var]] > 6), j = second_suit_var, value = NA)
+    set(vri_bem, i = which(vri_bem[[third_suit_var]] > 6), j = third_suit_var, value = NA)
 
-  set(vri_bem, i = which_su_1_is_na, j = "MURAR_PEFD_SU_1", value = 9)
-  set(vri_bem, i = which_su_2_is_na, j = "MURAR_PEFD_SU_2", value = 9)
-  set(vri_bem, i = which_su_3_is_na, j = "MURAR_PEFD_SU_3", value = 9)
+    which_na_list[[first_suit_var]] <- which(is.na(vri_bem[[first_suit_var]]))
+    which_na_list[[second_suit_var]] <- which(is.na(vri_bem[[second_suit_var]]))
+    which_na_list[[third_suit_var]] <- which(is.na(vri_bem[[third_suit_var]]))
 
+    set(vri_bem, i = which_na_list[[first_suit_var]], j = first_suit_var, value = 9)
+    set(vri_bem, i = which_na_list[[second_suit_var]], j = second_suit_var, value = 9)
+    set(vri_bem, i = which_na_list[[third_suit_var]], j = third_suit_var, value = 9)
 
-  vri_bem[ , MURAR_PEFD_SU_HV := fcase((MURAR_PEFD_SU_1 <= MURAR_PEFD_SU_2) & (MURAR_PEFD_SU_1 <= MURAR_PEFD_SU_3), MURAR_PEFD_SU_1,
-                                       MURAR_PEFD_SU_2 <= MURAR_PEFD_SU_3, MURAR_PEFD_SU_2,
-                                       MURAR_PEFD_SU_3 <= MURAR_PEFD_SU_2, MURAR_PEFD_SU_3,
-                                       default = NA)]
+    # calc best rating
+    fcase_expr <- parse_expr(paste0("fcase((",first_suit_var," <= ", second_suit_var, ") & (", first_suit_var, " <= ", third_suit_var, "), ", first_suit_var, ",
+                                            ", second_suit_var, " <= ", third_suit_var, ", ", second_suit_var, ",
+                                            ", third_suit_var, " <= ", second_suit_var, ", ", third_suit_var, ",
+                                            default = NA)"))
+    vri_bem[, (high_value_suit_var) := eval(fcase_expr)]
 
-   vri_bem[MURAR_PEFD_SU_HV > 8, MURAR_PEFD_SU_HV := NA]
+    set(vri_bem, i = which(vri_bem[[high_value_suit_var]] > 8), j = high_value_suit_var, value = NA)
 
-   # calc weighted suitability rating ----
+    # calc weighted suitability rating ----
 
-   # don't consider a rating that had not match in the RRM output but has percentage > 0
-   set(vri_bem, i = which_su_1_is_na, j = "MURAR_PEFD_SU_1", value = 0)
-   set(vri_bem, i = which_su_2_is_na, j = "MURAR_PEFD_SU_2", value = 0)
-   set(vri_bem, i = which_su_3_is_na, j = "MURAR_PEFD_SU_3", value = 0)
+    # don't consider a rating that had not match in the RRM output but has percentage > 0
+    set(vri_bem, i = which_na_list[[first_suit_var]], j = first_suit_var, value = 0)
+    set(vri_bem, i = which_na_list[[second_suit_var]], j = second_suit_var, value = 0)
+    set(vri_bem, i = which_na_list[[third_suit_var]], j = third_suit_var, value = 0)
 
-  vri_bem[, MURAR_PEFD_SU_WA := round(((MURAR_PEFD_SU_1 * SDEC_1) + (MURAR_PEFD_SU_2 * SDEC_2) + (MURAR_PEFD_SU_3 * SDEC_3))/(SDEC_1 * !is.na(MURAR_PEFD_SU_1) + SDEC_2 * !is.na(MURAR_PEFD_SU_2) + SDEC_3 * !is.na(MURAR_PEFD_SU_3)))]
-  vri_bem[MURAR_PEFD_SU_WA > 6 | MURAR_PEFD_SU_WA == 0, MURAR_PEFD_SU_WA := NA]
+    wa_expr <- parse_expr(paste0("round(((", first_suit_var, " * SDEC_1) + (", second_suit_var, " * SDEC_2) + (", third_suit_var, "* SDEC_3))/(SDEC_1 * !is.na(", first_suit_var, ") + SDEC_2 * !is.na(", second_suit_var, ") + SDEC_3 * !is.na(", third_suit_var, ")))"))
+    vri_bem[, (weighted_average_suit_var) := eval(wa_expr)]
+    set(vri_bem, i = which(vri_bem[[weighted_average_suit_var]] > 6 | vri_bem[[weighted_average_suit_var]] == 0 | is.nan(vri_bem[[weighted_average_suit_var]])), j = weighted_average_suit_var, value = NA)
+
+  }
 
   # calc highest capability value ----
 
-  # assign temporary worst rating  to rating NA to make calculation of best rating easier
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_CAP_1"]] > 6), j = "MURAR_PEFD_CAP_1", value = NA)
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_CAP_2"]] > 6), j = "MURAR_PEFD_CAP_2", value = NA)
-  set(vri_bem, i = which(vri_bem[["MURAR_PEFD_CAP_3"]] > 6), j = "MURAR_PEFD_CAP_3", value = NA)
+  for (cap_variable in cap_rating_variables) {
+    first_cap_var <- paste0(cap_variable, "_1")
+    second_cap_var <- paste0(cap_variable, "_2")
+    third_cap_var <- paste0(cap_variable, "_3")
+    high_value_cap_var <- paste0(cap_variable, "_HV")
+    weighted_average_cap_var <- paste0(cap_variable, "_WA")
 
-  which_cap_1_is_na <- which(is.na(vri_bem[["MURAR_PEFD_CAP_1"]]))
-  which_cap_2_is_na <- which(is.na(vri_bem[["MURAR_PEFD_CAP_2"]]))
-  which_cap_3_is_na <- which(is.na(vri_bem[["MURAR_PEFD_CAP_3"]]))
+    # assign temporary worst rating  to rating NA to make calculation of best rating easier
+    set(vri_bem, i = which(vri_bem[[first_cap_var]] > 6), j = first_cap_var, value = NA)
+    set(vri_bem, i = which(vri_bem[[second_cap_var]] > 6), j = second_cap_var, value = NA)
+    set(vri_bem, i = which(vri_bem[[third_cap_var]] > 6), j = third_cap_var, value = NA)
 
-  set(vri_bem, i = which_cap_1_is_na, j = "MURAR_PEFD_CAP_1", value = 9)
-  set(vri_bem, i = which_cap_2_is_na, j = "MURAR_PEFD_CAP_2", value = 9)
-  set(vri_bem, i = which_cap_3_is_na, j = "MURAR_PEFD_CAP_3", value = 9)
+    which_na_list[[first_cap_var]] <- which(is.na(vri_bem[[first_cap_var]]))
+    which_na_list[[second_cap_var]] <- which(is.na(vri_bem[[second_cap_var]]))
+    which_na_list[[third_cap_var]] <- which(is.na(vri_bem[[third_cap_var]]))
 
-  vri_bem[ , MURAR_PEFD_CAP_HV := fcase((MURAR_PEFD_CAP_1 <= MURAR_PEFD_CAP_2) & (MURAR_PEFD_CAP_1 <= MURAR_PEFD_CAP_3), MURAR_PEFD_CAP_1,
-                                       MURAR_PEFD_CAP_2 <= MURAR_PEFD_CAP_3, MURAR_PEFD_CAP_2,
-                                       MURAR_PEFD_CAP_3 <= MURAR_PEFD_CAP_2, MURAR_PEFD_CAP_3,
-                                       default = NA)]
-  vri_bem[MURAR_PEFD_CAP_HV > 8, MURAR_PEFD_CAP_HV := NA]
+    set(vri_bem, i = which_na_list[[first_cap_var]], j = first_cap_var, value = 9)
+    set(vri_bem, i = which_na_list[[second_cap_var]], j = second_cap_var, value = 9)
+    set(vri_bem, i = which_na_list[[third_cap_var]], j = third_cap_var, value = 9)
 
-  # calc weighted capability rating ----
+    # calc best rating
+    fcase_expr <- parse_expr(paste0("fcase((",first_cap_var," <= ", second_cap_var, ") & (", first_cap_var, " <= ", third_cap_var, "), ", first_cap_var, ",
+                                            ", second_cap_var, " <= ", third_cap_var, ", ", second_cap_var, ",
+                                            ", third_cap_var, " <= ", second_cap_var, ", ", third_cap_var, ",
+                                            default = NA)"))
+    vri_bem[, (high_value_cap_var) := eval(fcase_expr)]
 
-  vri_bem[, MURAR_PEFD_CAP_WA := round(((MURAR_PEFD_CAP_1 * SDEC_1) + (MURAR_PEFD_CAP_2 * SDEC_2) + (MURAR_PEFD_CAP_3 * SDEC_3))/(SDEC_1 + SDEC_2 + SDEC_3))]
-  vri_bem[MURAR_PEFD_CAP_WA > 6, MURAR_PEFD_CAP_WA := NA]
+    set(vri_bem, i = which(vri_bem[[high_value_cap_var]] > 8), j = high_value_cap_var, value = NA)
+
+    # calc weighted capability rating ----
+
+    # don't consider a rating that had not match in the RRM output but has percentage > 0
+    set(vri_bem, i = which_na_list[[first_cap_var]], j = first_cap_var, value = 0)
+    set(vri_bem, i = which_na_list[[second_cap_var]], j = second_cap_var, value = 0)
+    set(vri_bem, i = which_na_list[[third_cap_var]], j = third_cap_var, value = 0)
+
+    wa_expr <- parse_expr(paste0("round(((", first_cap_var, " * SDEC_1) + (", second_cap_var, " * SDEC_2) + (", third_cap_var, "* SDEC_3))/(SDEC_1 + SDEC_2 + SDEC_3))"))
+    vri_bem[, (weighted_average_cap_var) := eval(wa_expr)]
+    set(vri_bem, i = which(vri_bem[[weighted_average_cap_var]] > 6), j = weighted_average_cap_var, value = NA)
+
+  }
 
   # rating for decile that are 0 should be NA
-  set(vri_bem, i = which(vri_bem$SDEC_1 == 0), j = c("MURAR_PEFD_SU_1","MURAR_PEFD_CAP_1"), value = NA)
-  set(vri_bem, i = which(vri_bem$SDEC_2 == 0), j = c("MURAR_PEFD_SU_2","MURAR_PEFD_CAP_2"), value = NA)
-  set(vri_bem, i = which(vri_bem$SDEC_3 == 0), j = c("MURAR_PEFD_SU_3","MURAR_PEFD_CAP_3"), value = NA)
+  set(vri_bem, i = which(vri_bem$SDEC_1 == 0), j = first_decile_variables, value = NA)
+  set(vri_bem, i = which(vri_bem$SDEC_2 == 0), j = second_decile_variables, value = NA)
+  set(vri_bem, i = which(vri_bem$SDEC_3 == 0), j = third_decile_variables, value = NA)
 
-  set(vri_bem, i = which_su_1_is_na, j = "MURAR_PEFD_SU_1", value = NA)
-  set(vri_bem, i = which_su_2_is_na, j = "MURAR_PEFD_SU_2", value = NA)
-  set(vri_bem, i = which_su_3_is_na, j = "MURAR_PEFD_SU_3", value = NA)
-
-  set(vri_bem, i = which_cap_1_is_na, j = "MURAR_PEFD_CAP_1", value = NA)
-  set(vri_bem, i = which_cap_2_is_na, j = "MURAR_PEFD_CAP_2", value = NA)
-  set(vri_bem, i = which_cap_3_is_na, j = "MURAR_PEFD_CAP_3", value = NA)
+  for (rating_variable in names(which_na_list)) {
+    set(vri_bem, i = which_na_list[[rating_variable]], j = rating_variable, value = NA)
+  }
 
   set(vri_bem , j = "STAND_A1_temp", value = NULL)
   set(vri_bem , j = "STAND_A2_temp", value = NULL)
@@ -147,17 +180,82 @@ merge_rrm_on_vri_bear <- function(vri_bem, rrm_dt, return_sf = TRUE) {
 
 
 
-#' merge_rrm_on_vri_moose
+#' merge_rating_bear
 #'
 #' merge suitability and capability rating from rrm onto vri-bem
 #' and calculate highest value and weighted average for each scores
 #'
 #' @param vri_bem sf object that represent VRI (vegetation ressource inventory) features
 #' @param rrm_dt data.table object that contains the rrm
-#' @param return_sf logical, if TRUE  return sf object , if FALSE return data.table object and update by reference
+#' @param rating_variables list, vector for each rating variables to create
+#' @param rating_variable_expr parse_expr , expression for variable to merge in data.table
 #' @return sf  vri-bem object with new columns for rating
 #' @import data.table
 #' @export
-merge_rrm_on_vri_moose <- function(vri_bem, rrm_dt, return_sf = TRUE) {
+merge_rating_bear <- function(vri_bem, rrm_dt, rating_variables, rating_variables_expr) {
+  # merge on decile 1 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S1 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_BEAR_1 = Crown_Bear, STRCT_S1 = Strct_d, STAND_A1_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[1]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
+  # merge on decile 2 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S2 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_BEAR_2 = Crown_Bear, STRCT_S2 = Strct_d, STAND_A2_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[2]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
+
+
+  # merge on decile 3 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S3 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_BEAR_3 = Crown_Bear, STRCT_S3 = Strct_d, STAND_A3_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[3]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
+
+  return(vri_bem)
+}
+
+
+#' merge_rating_moose
+#'
+#' merge suitability and capability rating from rrm onto vri-bem
+#' and calculate highest value and weighted average for each scores
+#'
+#' @param vri_bem sf object that represent VRI (vegetation ressource inventory) features
+#' @param rrm_dt data.table object that contains the rrm
+#' @param rating_variables list, vector for each rating variables to create
+#' @param rating_variable_expr parse_expr , expression for variable to merge in data.table
+#' @return sf  vri-bem object with new columns for rating
+#' @import data.table
+#' @export
+merge_rating_moose <- function(vri_bem, rrm_dt, rating_variables, rating_variables_expr) {
+  # merge on decile 1 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S1 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_MOOSE_1 = Crown_Moose, STRCT_S1 = Strct_d, STAND_A1_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[1]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
+  # merge on decile 2 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S2 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_MOOSE_2 = Crown_Moose, STRCT_S2 = Strct_d, STAND_A2_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[2]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
+
+
+  # merge on decile 3 ----
+  eval(parse_expr(paste0("vri_bem[rrm_dt, on = .(ECO_SEC = Eco_sec, BGC_ZONE = Bgc_zone, BGC_SUBZON = Bgc_subzon, BGC_VRT = Bgc_vrt,
+                         BGC_PHASE = Bgc_phase, BEUMC_S3 = Beumc, SLOPE_MOD = Slope_mod,
+                         SITE_M3A = Site_m3a, SNOW_CODE = Snow_code, ABOVE_ELEV = Above_Elev_Thold,
+                         CROWN_MOOSE_3 = Crown_Moose, STRCT_S3 = Strct_d, STAND_A3_temp = Stand_d),
+          c(", paste0("'", paste(rating_variables[[3]], collapse = "','"), "'"), ") := ", rating_variables_expr,"]")))
+
   return(vri_bem)
 }